requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
motor>=3.3.0
python-dotenv>=1.0.0
websockets>=13.0
